加了底部导航栏的组件
